Refactor: 버튼 경로 수정
